+++ docs/source/changelog.rst
@@ -4,8 +4,11 @@
 Changelog
 =========
 
+- :release:`1.8.0 <11th March 2023>`
+- :feature:`26` Move random number generation to generators module
+
 - :release:`1.7.0 <11th March 2023>`
-- :feature:`24` Add random number generation
+- :feature:`24` Add bulk uuid generation
 
 - :release:`1.6.0 <09th March 2023>`
 - :feature:`21` Add random number generation